refactor: AcceptanceTestBase와 ResponseParser를 분리함
