+++ src/test/java/nextstep/subway/acceptance/AcceptanceTestBase.java
@@ -14,16 +14,4 @@ public class AcceptanceTestBase {
         assertThat(response.statusCode()).isEqualTo(httpStatus.value());
     }
 
-    public static Long getIdFromResponse(final ExtractableResponse<Response> response) {
-        return response.jsonPath().getLong("id");
-    }
-
-    public static String getNameFromResponse(final ExtractableResponse<Response> response) {
-        return response.jsonPath().getString("name");
-    }
-
-    public static List<String> getNamesFromResponse(final ExtractableResponse<Response> response) {
-        return response.jsonPath().getList("name", String.class);
-    }
-
 }
