+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -15,7 +15,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static nextstep.subway.acceptance.AcceptanceTestBase.*;
+import static nextstep.subway.acceptance.AcceptanceTestBase.assertStatusCode;
+import static nextstep.subway.acceptance.ResponseParser.getIdFromResponse;
+import static nextstep.subway.acceptance.ResponseParser.getNamesFromResponse;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철역 관련 기능")
